DAO code and logger code has been updated

+++ restaurent_registration/src/main/java/com/gogenie/restaurent/registration/dao/RestaurentRegistraionDAO.java
@@ -9,5 +9,5 @@ public interface RestaurentRegistraionDAO {
 	
 	public String updateExistingRestaurentDetails(RestaurentRegistrationRequest request) throws RestaurentRegistrationException;
 	
-	public String acivateOrDeactiveARestaurent(Integer restaurentId, String restaurentName, String isActiveFlag) throws RestaurentRegistrationException;
+	public String acivateOrDeactiveARestaurent(Long restaurentId, String restaurentName, String isActiveFlag) throws RestaurentRegistrationException;
 }
