+++ restaurent_registration/src/main/java/com/gogenie/restaurent/registration/service/RestaurentRegistraionService.java
@@ -6,7 +6,10 @@ import com.gogenie.restaurent.registration.model.RestaurentRegistrationRequest;
 public interface RestaurentRegistraionService {
 
 	public String registerRestaurent(RestaurentRegistrationRequest request) throws RestaurentRegistrationException;
+
 	public String updateRestaurentDetails(RestaurentRegistrationRequest request) throws RestaurentRegistrationException;
-	public String activateAndDeactivateARestaurent(Integer restaurentId, String restaurentName, String isActiveFlag) throws RestaurentRegistrationException;
-	
+
+	public String activateAndDeactivateARestaurent(Long restaurentId, String restaurentName, String isActiveFlag)
+			throws RestaurentRegistrationException;
+
 }
